% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govcan_dl_resources.R
\name{govcan_dl_resources}
\alias{govcan_dl_resources}
\title{Download the resources attached to a specific record or (i.e. a CKAN package)
or to a stack of packages,}
\usage{
govcan_dl_resources(resources, ...)
}
\arguments{
\item{resources}{An object of the class ckan_package_stack or ckan_package, or an id of
a specific record or (i.e. a CKAN package), or an object of type ckan_resource or
ckan_resource_stack}

\item{file_formats}{(character vector) A character vector with file formats to be
donwloaded, any of :
\enumerate{
  \item CSV
  \item JSON
  \item SHP
}}

\item{where}{(string) One of "session" is files have to be charged in the session
or a path to the folder in which to download the files}
}
\description{
Download resources attached to a specific record or (i.e. a CKAN package)
or to a stack of packages
}
