% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govcan_search.R
\name{govcan_search}
\alias{govcan_search}
\title{Query OpenCan portal for datasets matching keywords}
\usage{
govcan_search(keywords, records = 10, only_results = TRUE,
  format_results = TRUE, ... = NULL)
}
\arguments{
\item{keywords}{(character) A set of keywords to query}

\item{records}{(numeric) The number of matching records to return from the CKAn query
(number of rows in the JSON output)}

\item{only_results}{(logical) Whether the function should return only the results
(default, recommanded to be able to use other functions in the package)}

\item{format_results}{(logical) Whether the function should return a formatted output
of the results as a tibble (default) or an unformatted version under the form of
CKAN packages contaning lists}

\item{...}{More arguments to be passed on to ckanr::package_search()}
}
\value{
If only_results is TRUE, will return only the results of the search, else it
will return all the output of the CKAN query. If format_results is TRUE, the results
are formatted to a tibble.
}
\description{
This function wraps ckanr::package_search() to search for any records
matching a given set of jeywords within the open Canada portal
}
