% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govcan_setup.R
\name{govcan_setup}
\alias{govcan_setup}
\title{Set or reset the ckanr URL to the Open Canada portal URl "https://open.canada.ca/data/en"}
\usage{
govcan_setup(url = "https://open.canada.ca/data/en")
}
\description{
Set or reset the ckanr URL to the Open Canada portal URl
"https://open.canada.ca/data/en"
}
