% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govcan_show_resources.R
\name{govcan_show_resources}
\alias{govcan_show_resources}
\title{Display resources attached to a specific record or (i.e. a CKAN package) or to a
stack of packages}
\usage{
govcan_show_resources(x)
}
\arguments{
\item{x}{An object of the class ckan_package_stack or ckan_package}
}
\value{
If format is TRUE, will return a tibble of available file formats for a
given datset or query, else it will return an equivalent list (default).
}
\description{
Display resources attached to a specific record or (i.e. a CKAN package) or to a
stack of packages
}
