% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckan_resource_stack-class.R
\name{new_ckan_resource_stack}
\alias{new_ckan_resource_stack}
\title{ckan_resource_stack class and helpers}
\usage{
new_ckan_resource_stack(x = list())
}
\arguments{
\item{x}{a list}
}
\description{
ckan_resource_stack class and helpers
}
\keyword{internal}
