% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencan_get_resources.R
\name{opencan_get_ressources}
\alias{opencan_get_ressources}
\title{Query OpenCan portal for a specific dataset}
\usage{
opencan_get_ressources(record_id = NULL, query_results = NULL,
  domain = "https://open.canada.ca/data/en/api/3/")
}
\arguments{
\item{record_id}{Id of the wanated dataset, of the form
"4a2929ce-d6b1-49b0-b520-63be0859c552"}

\item{query_results}{The results of a call to `opencan_search`}

\item{domain}{Domain name for the OpenCan API being querried}
}
\value{
If format is TRUE, will return a dataframe of available file formats for a
given datset or query, else it will return the unformatted json R object from
jsonlite::fromJSON
}
\description{
Return

One of @param record_id or @param query_results
}
