% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govcan_setup.R
\name{govcan_setup}
\alias{govcan_setup}
\title{Set or reset the ckanr URL}
\usage{
govcan_setup(url = "https://open.canada.ca/data/en")
}
\arguments{
\item{url}{Open Canada portal URL.}
}
\value{
This function invisibly returns the URL sent as input.
}
\description{
Set or reset the ckanr URL to the Open Canada portal URL
\url{https://open.canada.ca/data/en}
}
\examples{
\dontrun{
govcan_setup("https://open.canada.ca/data/en")
}

}
