% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencan_search.R
\name{opencan_search}
\alias{opencan_search}
\title{Query OpenCan portal for datasets matching keywords}
\usage{
opencan_search(keywords,
  domain = "https://open.canada.ca/data/en/api/3/", format = TRUE)
}
\arguments{
\item{keywords}{A set of keywords to query}

\item{domain}{Domain name for the OpenCan API being querried}

\item{format}{Whether the function should return a formatted output}
}
\description{
Query OpenCan portal for datasets matching keywords
}
